modify all regressors to make use of Keras-based kwargs
