+++ sktime/regression/deep_learning/base.py
@@ -4,13 +4,14 @@ The reason for this class between BaseClassifier and deep_learning classifiers i
 because we can generalise tags and _predict
 """
 
-__all__ = ["BaseDeepRegressor"]
+__all__ = ["BaseDeepRegressor", "KerasCompileKwargs", "KerasFitKwargs"]
 
 import os
 from abc import abstractmethod
 
 import numpy as np
 
+from sktime.base._base_tf import KerasCompileKwargs, KerasFitKwargs
 from sktime.regression.base import BaseRegressor
 from sktime.utils.dependencies import _check_soft_dependencies
 
@@ -24,8 +25,12 @@ class BaseDeepRegressor(BaseRegressor):
 
     Parameters
     ----------
-    batch_size : int, default = 40
-        training batch size for the model
+    compile_kwargs : KerasCompileKwargs, default=None
+        Additional arguments for Keras model compilation.
+        See ``KerasCompileKwargs`` for available options.
+    fit_kwargs : KerasFitKwargs, default=None
+        Additional arguments for Keras model training.
+        See ``KerasFitKwargs`` for available options.
 
     Attributes
     ----------
@@ -40,6 +45,16 @@ class BaseDeepRegressor(BaseRegressor):
         "property:randomness": "stochastic",
         "capability:random_state": True,
     }
+
+    def __init__(
+        self,
+        compile_kwargs: KerasCompileKwargs | None = None,
+        fit_kwargs: KerasFitKwargs | None = None,
+    ):
+        self.compile_kwargs = compile_kwargs
+        self.fit_kwargs = fit_kwargs
+
+        super().__init__()
 
     @abstractmethod
     def build_model(self, input_shape, **kwargs):
