+++ sktime/regression/deep_learning/lstmfcn.py
@@ -54,6 +54,12 @@ class LSTMFCNRegressor(BaseDeepRegressor):
         Activation function used in the hidden layers.
         List of available activation functions:
         https://keras.io/api/layers/activations/
+    compile_kwargs : KerasCompileKwargs, default=None
+        Additional arguments for Keras model compilation.
+        See ``KerasCompileKwargs`` for available options.
+    fit_kwargs : KerasFitKwargs, default=None
+        Additional arguments for Keras model training.
+        See ``KerasFitKwargs`` for available options.
 
     References
     ----------
@@ -95,6 +101,8 @@ class LSTMFCNRegressor(BaseDeepRegressor):
         verbose=0,
         activation="linear",
         activation_hidden="relu",
+        compile_kwargs=None,
+        fit_kwargs=None,
     ):
         # predefined
         self.activation = activation
@@ -111,7 +119,7 @@ class LSTMFCNRegressor(BaseDeepRegressor):
         self.random_state = random_state
         self.verbose = verbose
 
-        super().__init__()
+        super().__init__(compile_kwargs=compile_kwargs, fit_kwargs=fit_kwargs)
 
         self.input_shape = None
         self.model_ = None
@@ -153,10 +161,14 @@ class LSTMFCNRegressor(BaseDeepRegressor):
 
         model = keras.models.Model(inputs=input_layers, outputs=output_layer)
 
+        compile_kwargs = (
+            self.compile_kwargs.as_dict() if self.compile_kwargs is not None else {}
+        )
         model.compile(
             loss="mean_squared_error",
             optimizer="sgd",
             metrics=["accuracy"],
+            **compile_kwargs,
         )
 
         self._callbacks = self.callbacks or None
@@ -194,6 +206,7 @@ class LSTMFCNRegressor(BaseDeepRegressor):
         if self.verbose:
             self.model_.summary()
 
+        fit_kwargs = self.fit_kwargs.as_dict() if self.fit_kwargs is not None else {}
         self.history = self.model_.fit(
             X,
             y,
@@ -201,6 +214,7 @@ class LSTMFCNRegressor(BaseDeepRegressor):
             epochs=self.n_epochs,
             verbose=self.verbose,
             callbacks=deepcopy(self._callbacks) if self._callbacks else None,
+            **fit_kwargs,
         )
 
         return self
