+++ sktime/regression/deep_learning/mcdcnn.py
@@ -62,6 +62,12 @@ class MCDCNNRegressor(BaseDeepRegressor):
         The callback(s) to use during training.
     random_state : int, optional (default=0)
         The seed to any random action.
+    compile_kwargs : KerasCompileKwargs, default=None
+        Additional arguments for Keras model compilation.
+        See ``KerasCompileKwargs`` for available options.
+    fit_kwargs : KerasFitKwargs, default=None
+        Additional arguments for Keras model training.
+        See ``KerasFitKwargs`` for available options.
 
     References
     ----------
@@ -106,6 +112,8 @@ class MCDCNNRegressor(BaseDeepRegressor):
         optimizer=None,
         verbose=False,
         random_state=0,
+        compile_kwargs=None,
+        fit_kwargs=None,
     ):
         _check_dl_dependencies(severity="error")
 
@@ -127,7 +135,7 @@ class MCDCNNRegressor(BaseDeepRegressor):
         self.verbose = verbose
         self.random_state = random_state
 
-        super().__init__()
+        super().__init__(compile_kwargs=compile_kwargs, fit_kwargs=fit_kwargs)
 
         self.history = None
         self._network = MCDCNNNetwork(
@@ -184,10 +192,14 @@ class MCDCNNRegressor(BaseDeepRegressor):
 
         model = keras.models.Model(inputs=input_layers, outputs=output_layer)
 
+        compile_kwargs = (
+            self.compile_kwargs.as_dict() if self.compile_kwargs is not None else {}
+        )
         model.compile(
             loss=self.loss,
             optimizer=self.optimizer_,
             metrics=metrics,
+            **compile_kwargs,
         )
 
         return model
@@ -218,12 +230,14 @@ class MCDCNNRegressor(BaseDeepRegressor):
         if self.verbose:
             self.model_.summary()
 
+        fit_kwargs = self.fit_kwargs.as_dict() if self.fit_kwargs is not None else {}
         self.history = self.model_.fit(
             X,
             y,
             batch_size=self.batch_size,
             verbose=self.verbose,
             callbacks=self.callbacks_,
+            **fit_kwargs,
         )
 
         return self
